requirements.txt
# GitHub Agent V2 - Requirements
# Phase 1: Core Framework

# Web Framework
flask>=2.3.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.8.0
pathlib>=1.0.1

# Phase 2: Cloud Agent (OpenClaw)
# openclaw-cli  # 如果 OpenClaw 有 pip 包

# Phase 3: Knowledge Base
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
# numpy==1.26.4  # 固定在 1.26.4 避免兼容性问题

# Phase 4: Code Executor
# 使用本地 Ollama，无需额外依赖

# Phase 5: Infrastructure
# Configuration (optional, for better validation)
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
